Allow configuration of datacenter.cfg
Add a variable to set configuration options in the
/etc/pve/datacenter.cfg configuration file.

https://pve.proxmox.com/wiki/Manual:_datacenter.cfg

The datacenter.cfg file has to exist before it can be successfully
written by the Ansible copy module. This is because the pmxcfs used for
/etc/pve only has very limited support for setting file permissions and
ownership. The copy module tries to set unsupported permissions in an
intermediate step if the file does not yet exist. This is worked around
by creating the file using the Ansible file module if it does not yet
exist.