Don't modprobe in Travis CI tests
A change in the modprobe module in Ansible 2.8 made the module fail even
if the state is already correct. Because of this the task "Unload
nmi_watchdog" failed in tests.

Loading and unloading modules does not work when running tests on Travis
CI in an LXC container. This adds the "skiponlxc" tag to all invocations of
the modprobe Ansible module.