change acl "propagate" attribute to integer
The `get access/acl` call returns `propagate` as an integer value. However the module was defining it as a string. This resulted in the module constantly redefining the ACL and reporting it as "changed" when nothing was actually changing.

Not sure if at some point in the past this actually was a string, but as of 6.3 it is an integer.