Only update SSH known_hosts when joining cluster
This role doesn't need to add all cluster hosts to /etc/ssh/ssh_known_hosts as
PVE's cluster setup tool does this for you when joining a cluster. However, it
is necessary for the node joining a cluster to already know the public host
keys of the node it is using to join a cluster, so we add it temporarily to the
root user's known_hosts. We don't add it to /etc/ssh/ssh_known_hosts as `pvecm`
merges the contents of it into the cluster's copy of it, which is undesirable
(it adds extraneous comments from ansible).