Fix reinstall of initialization node
By default, the initialization node is the first node in the group
`pve_group`. When this node is reinstalled, it is not possible for the
role to wait for quorum on this node since it is not a member of the
cluster yet. Moreover, the tasks in pve_add_node.yml are not run for
this node.

Change the behavior of the role to pick any node already configured in
`pve_cluster_clustername` and use this one as an initialization node.

This does not change the behavior when installing a new cluster.