Fix host used to configure HA groups
When the role is run with `--limit nodes_without_first_node`, the HA
groups are not configured (tasks are skipped). Use `_init_node` instead
of `groups[pve_group][0]` to run the tasks.

`_init_node` is any active node in an already initialized Proxmox
cluster which is part of the run.